Updated Simulator Setup to work with Sanity

+++ scripts/coprhd_cli.sh
@@ -1,4 +1,4 @@
-#! /bin/bash
+#!/bin/bash
 #################################################
 # Install CoprHD CLI Scripts
 #################################################
@@ -16,6 +16,14 @@ do
       ;;
     -s|--simulator)
       simulator="$2"
+      shift
+      ;;
+    -a|--all_simulators)
+      all_simulators="$2"
+      shift
+      ;;
+    -i|--node_ip)
+      coprhd_ip="$2"
       shift
       ;;
     *)
@@ -89,15 +97,18 @@ group_member_attributes:member,roleOccupant,memberUid",uniqueMember
 group_object_classes:posixGroup,organizationalRole,groupOfNames,groupOfUniqueNames
 EOF2
 
-# Grab the SMIS Simulator, if desired
-if [ "$simulator" = true ]; then
+# Grab the SMIS Simulator if enabled
+SIMULATOR_VERSION="smis-simulators-1.0.0.0.1455598800.zip"
+if [ "$simulator" = true ] || [ "$all_simulators" = true ]; then
   # Download to /vagrant directory if needed
-  if [ ! -e /vagrant/smis_simulator.zip ]; then
-     wget 'https://coprhd.atlassian.net/wiki/download/attachments/6652057/smis-simulator.zip?version=2&modificationDate=1444855261258&api=v2' -O /vagrant/smis_simulator.zip
-    wget 'https://coprhd.atlassian.net/wiki/download/attachments/6652057/cisco-sim.zip?version=4&modificationDate=1453406325249&api=v2 -O /vagrant/cisco_sim.zip'
+  if [ ! -e /vagrant/$SIMULATOR_VERSION ]; then
+     wget "https://coprhd.atlassian.net/wiki/download/attachments/6652057/$SIMULATOR_VERSION?version=1&modificationDate=1455833007237&api=v2" -O "/vagrant/$SIMULATOR_VERSION"
   fi
   # Install SMIS
-  unzip /vagrant/smis_simulator.zip -d /opt/storageos/
+  unzip /vagrant/$SIMULATOR_VERSION -d /opt/storageos/
+  # Enable version 4.6.2 SMI-S for Sanity Testing
+  sed -i 's/^VERSION=80/#VERSION=80/' /opt/storageos/ecom/providers/OSLSProvider.conf
+  sed -i 's/^#VERSION=462/VERSION=462/' /opt/storageos/ecom/providers/OSLSProvider.conf 
   cd /opt/storageos/ecom/bin
   chmod +x  ECOM
   chmod +x  system/ECOM
@@ -111,15 +122,78 @@ if [ "$simulator" = true ]; then
     printf "."
     sleep $INTERVAL
   done
-  # Install Cisco Sim
+fi  # SMIS_Simulator
+
+# Grab All Simulators and Install (SMIS already done above)
+if [ "$all_simulators" = true ]; then
+  echo "Installing all Simulators"
+  # First Cisco Simulator
   mkdir /simulator
-  mv /vagrant/cisco_sim.zip /simulator
+  wget 'https://coprhd.atlassian.net/wiki/download/attachments/6652057/cisco-sim.zip?version=4&modificationDate=1453406325249&api=v2' -O /simulator/cisco_sim.zip
   cd /simulator
   unzip cisco_sim.zip
   cd cisco-sim
-  cp bashrc ~/.bashrc 
   # Update Config files for correct directory
+  cp bashrc ~/.bashrc 
   sed -i 's/CISCO_SIM_HOME=\/cisco-sim/CISCO_SIM_HOME=\/simulator\/cisco-sim/' ~/.bashrc
   sed -i 's/chmod -R 777 \/cisco-sim/chmod -R 777 \/simulator\/cisco-sim/' ~/.bashrc
-  sed -i 's#args=\(\'\/cisco-sim/args=\(\'\/simulator\/cisco-sim/# /simulator/cisco-sim/config/logging.conf
-fi
+  source ~/.bashrc
+  sed -i "s#args=('\/cisco-sim\/#args=('\/simulator\/cisco-sim\/#" /simulator/cisco-sim/config/logging.conf 
+  # Update sshd_config to allow root login - that's how Cisco Sim works
+  sed -i "s/PermitRootLogin no/PermitRootLogin yes/" /etc/ssh/sshd_config
+  service sshd restart
+
+  # Second, LDAP Simulator
+  wget 'https://coprhd.atlassian.net/wiki/download/attachments/6652057/ldapsvc-1.0.0.zip?version=2&modificationDate=1453406325338&api=v2' -O /simulator/ldap.zip
+  cd /simulator
+  unzip ldap.zip
+  cd /simulator/ldapsvc-1.0.0/bin/
+  echo "Starting LDAP Simulator Service"
+  ./ldapsvc &
+  sleep 5
+  curl -X POST -H "Content-Type: application/json" -d "{\"listener_name\": \"COPRHDLDAPSanity\"}" http://${coprhd_ip}:8082/ldap-service/start
+
+  # Third, Windows Host Simulator
+  wget 'https://coprhd.atlassian.net/wiki/download/attachments/6652057/win-sim.zip?version=3&modificationDate=1453406324934&api=v2' -O /simulator/win_host.zip
+  cd /simulator
+  unzip win_host.zip
+  cd win-sim
+  # Update Provider IP for SMIS Simulator address (running on CoprHD in this setup)
+  sed -i "s/<provider ip=\"10.247.66.220\" username=\"admin\" password=\"#1Password\" port=\"5989\" type=\"VMAX\"><\/provider>/<provider ip=\"${coprhd_ip}\" username=\"admin\" password=\"#1Password\" port=\"5989\" type=\"VMAX\"><\/provider>/" /simulator/win-sim/config/simulator.xml
+   echo "${coprhd_ip} winhost1 winhost2 winhost3 winhost4 winhost5 winhost6 winhost7 winhost8 winhost9 winhost10" >> /etc/hosts
+  ./runWS.sh &
+  sleep 5
+
+  # Fourth, VPLEX Simulator
+  wget 'https://coprhd.atlassian.net/wiki/download/attachments/6652057/vplex-sim.zip?version=4&modificationDate=1453406325096&api=v2' -O /simulator/vplex.zip
+  cd /simulator
+  unzip vplex.zip
+  cd vplex-simulators-1.0.0.0.41/
+  # Edit IP Address for the SMIS provider and Vplex Simulator address (both CoprHD IP in this setup)
+  sed -i "s/SMIProviderIP=10.247.98.128:5989,10.247.98.128:7009/SMIProviderIP=${coprhd_ip}:5989/" vplex_config.properties
+  sed -i "s/#VplexSimulatorIP=10.247.98.128/VplexSimulatorIP=${coprhd_ip}/" vplex_config.properties
+  #sed -i 's/RP_ENABLE=true/RP_ENABLE=false/' vplex_config.properties
+  chmod +x ./run.sh
+  ./run.sh &
+  # Need to wait for service to be running
+  sleep 2
+  PID=`ps -ef | grep [v]plex_config | awk '{print $2}'`
+  if [[ -z ${PID} ]]; then
+     echo "Vplex_Config Simulator Not running - Fail"
+     exit 1
+  fi
+  TIMER=1
+  INTERVAL=3
+  echo "Waiting for VPlex Simulator to Start..."
+  while [[ "`netstat -anp | grep 4430 | grep -c ${PID}`" == 0 ]];
+    do
+      if [ $TIMER -gt 10 ]; then
+      echo ""
+      echo "VPlex Sim did not start!" >&2
+      exit 1
+    fi
+      printf "."
+      sleep $INTERVAL
+      let TIMER=TIMER+$INTERVAL
+    done
+fi # All_Simulators
